to_date: coerce argument to string
Since the value passed in can be an int (since it's coming from YAML),
we need to convert it to a string.

This was actually already in the code at some point, but I forgot why it
was there and removed it thinking it was superfluous.  I changed a test
so that it passes an int instead of the string as a year.